MdeModulePkg/Bds: Fix build failures of VS tool chain
Contributed-under: TianoCore Contribution Agreement 1.0

+++ MdeModulePkg/Library/UefiBootManagerLib/BmMisc.c
@@ -232,7 +232,7 @@ BmSetMemoryTypeInformationVariable (
     // Do not count the reserved memory occupied by RAM Disk.
     //
     if (CurrentMemoryTypeInformation[Index1].Type == EfiReservedMemoryType) {
-      CurrentMemoryTypeInformation[Index1].NumberOfPages -= RamDiskSizeInPages;
+      CurrentMemoryTypeInformation[Index1].NumberOfPages -= (UINT32) RamDiskSizeInPages;
     }
 
     //
